# doc/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: TRO & LAM API
  version: 1.5.0
  description: |
    Contract frozen from doc/planning/04-data-model.md and 05-api-contracts.md.
    All application routes use /api/v1. Health probes keep the scaffold response
    shape. Provider webhook wire payload and ACK are delegated to the current
    provider documentation; browser redirects never mutate payment state.
servers:
  - url: /api/v1
...
paths:
...
  /auth/csrf:
    get:
      operationId: getCsrfToken
      tags: [auth]
      x-actors: [guest, customer, staff, admin]
      responses: { '200': { $ref: '#/components/responses/Success' }, default: { $ref: '#/components/responses/Error' } }
  /auth/register:
    post:
      operationId: registerCustomer
      tags: [auth]
      x-actors: [guest]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserRegister' }
      responses: { '202': { $ref: '#/components/responses/Accepted' }, '400': { $ref: '#/components/responses/Error' }, '429': { $ref: '#/components/responses/Error' } }
  /auth/invitations/accept:
    post:
      operationId: acceptInvitation
      tags: [auth]
      x-actors: [guest]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/InvitationAccept' } } } }
      responses: { '200': { $ref: '#/components/responses/Success' }, '400': { $ref: '#/components/responses/Error' }, '410': { $ref: '#/components/responses/Error' }, '429': { $ref: '#/components/responses/Error' } }
  /auth/verify-email:
    post:
      operationId: verifyEmail
      tags: [auth]
      x-actors: [guest]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [token], properties: { token: { type: string } }, additionalProperties: false } } } }
      responses: { '200': { $ref: '#/components/responses/Success' }, '410': { $ref: '#/components/responses/Error' } }
  /auth/resend-verification:
    post:
      operationId: resendVerification
      tags: [auth]
      x-actors: [guest]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [email], properties: { email: { type: string, format: email } }, additionalProperties: false } } } }
      responses: { '202': { $ref: '#/components/responses/Accepted' }, '429': { $ref: '#/components/responses/Error' } }
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      x-actors: [guest]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [email, password], properties: { email: { type: string, format: email }, password: { type: string } }, additionalProperties: false } } } }
      responses: { '200': { $ref: '#/components/responses/Success' }, '403': { $ref: '#/components/responses/Error' }, '429': { $ref: '#/components/responses/Error' } }
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      x-actors: [guest, customer, staff, admin]
      responses: { '204': { $ref: '#/components/responses/NoContent' }, default: { $ref: '#/components/responses/Error' } }
  /auth/forgot-password:
    post:
      operationId: requestPasswordReset
      tags: [auth]
      x-actors: [guest]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [email], properties: { email: { type: string, format: email } }, additionalProperties: false } } } }
      responses: { '202': { $ref: '#/components/responses/Accepted' }, '429': { $ref: '#/components/responses/Error' } }
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      x-actors: [guest]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [token, password], properties: { token: { type: string }, password: { type: string, minLength: 12, maxLength: 128 } }, additionalProperties: false } } } }
      responses: { '200': { $ref: '#/components/responses/Success' }, '410': { $ref: '#/components/responses/Error' } }
  /account/password:
    post:
      operationId: changeOwnPassword
      tags: [account]
      x-actors: [customer, staff, admin]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ChangePassword' } } } }
      responses: { '200': { $ref: '#/components/responses/Success' }, '401': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
  /account/email-change:
    post:
      operationId: requestOwnEmailChange
      tags: [account]
      x-actors: [customer, staff, admin]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/EmailChange' } } } }
      responses: { '202': { $ref: '#/components/responses/Accepted' }, '401': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
  /account/email-change/verify:
    post:
      operationId: verifyOwnEmailChange
      tags: [account]
      x-actors: [customer, staff, admin]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/EmailChangeVerify' } } } }
      responses: { '200': { $ref: '#/components/responses/Success' }, '401': { $ref: '#/components/responses/Error' }, '403': { $ref: '#/components/responses/Error' }, '410': { $ref: '#/components/responses/Error' } }
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      x-actors: [customer, staff, admin]
      responses: { '200': { $ref: '#/components/responses/Success' }, '401': { $ref: '#/components/responses/Error' }, '403': { $ref: '#/components/responses/Error' } }
  /auth/appeal-challenges:
    post:
      operationId: createAppealChallenge
      tags: [auth]
      x-actors: [guest]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [email], properties: { email: { type: string, format: email } }, additionalProperties: false } } } }
      responses: { '202': { $ref: '#/components/responses/Accepted' }, '429': { $ref: '#/components/responses/Error' } }
  /auth/appeal-access:
    post:
      operationId: exchangeAppealAccess
      tags: [auth]
      x-actors: [guest]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [email, verificationCode, challengeId], properties: { email: { type: string, format: email }, verificationCode: { type: string }, challengeId: { type: string } }, additionalProperties: false } } } }
      responses: { '200': { $ref: '#/components/responses/Success' }, '403': { $ref: '#/components/responses/Error' }, '410': { $ref: '#/components/responses/Error' } }
  /account/appeals/current:
    get:
      operationId: getCurrentAppeal
      tags: [account]
      x-actors: [blocked-appeal]
      responses: { '200': { $ref: '#/components/responses/Success' }, '401': { $ref: '#/components/responses/Error' }, '403': { $ref: '#/components/responses/Error' } }
  /account/appeals:
    post:
      operationId: submitAppeal
      tags: [account]
      x-actors: [blocked-appeal]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [message], properties: { message: { type: string, minLength: 1, maxLength: 5000 } }, additionalProperties: false } } } }
      responses: { '201': { $ref: '#/components/responses/Created' }, '409': { $ref: '#/components/responses/Error' } }
  /account/profile:
    patch:
      operationId: updateProfile
      tags: [account]
      x-actors: [customer, staff, admin]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ProfileWrite' } } } }
      responses: { '200': { $ref: '#/components/responses/Success' }, '401': { $ref: '#/components/responses/Error' }, '403': { $ref: '#/components/responses/Error' } }
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      x-actors: [admin]
      parameters: [{ in: query, name: q, schema: { type: string, maxLength: 120 } }, { in: query, name: role, schema: { $ref: '#/components/schemas/Role' } }, { in: query, name: status, schema: { $ref: '#/components/schemas/UserStatus' } }, { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }]
      responses: { '200': { $ref: '#/components/responses/Success' }, '403': { $ref: '#/components/responses/Error' } }
    post:
      operationId: inviteAdminUser
      tags: [admin]
      x-actors: [admin]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UserInvite' } } } }
      responses: { '201': { $ref: '#/components/responses/Created' }, '403': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
  /admin/users/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      operationId: getAdminUser
      tags: [admin]
      x-actors: [admin]
      responses: { '200': { $ref: '#/components/responses/Success' }, '403': { $ref: '#/components/responses/Error' }, '404': { $ref: '#/components/responses/Error' } }
    patch:
      operationId: updateAdminUser
      tags: [admin]
      x-actors: [admin]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AdminUserWrite' } } } }
      responses: { '200': { $ref: '#/components/responses/Success' }, '403': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
  /admin/users/{id}/status:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      operationId: updateAdminUserStatus
      tags: [admin]
      x-actors: [admin]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UserStatusWrite' } } } }
      responses: { '200': { $ref: '#/components/responses/Success' }, '403': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
  /admin/users/{id}/role:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      operationId: updateAdminUserRole
      tags: [admin]
      x-actors: [admin]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UserRoleWrite' } } } }
      responses: { '200': { $ref: '#/components/responses/Success' }, '403': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
  /admin/users/{id}/password-reset:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      operationId: requestAdminPasswordReset
      tags: [admin]
      x-actors: [admin]
      description: Requires active admin capability users.manage and CSRF protection. Sends a one-time reset link only to the target account's verified email. The raw token is never returned. This request does not change the password or revoke sessions; the existing reset completion endpoint performs that transition. Persistent per-target cooldown applies.
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AdminPasswordResetRequest' } } } }
      responses: { '202': { $ref: '#/components/responses/Accepted' }, '403': { $ref: '#/components/responses/Error' }, '404': { $ref: '#/components/responses/Error' }, '429': { $ref: '#/components/responses/Error' }, '503': { $ref: '#/components/responses/Error' } }
  /admin/appeals:
    get:
      operationId: listAdminAppeals
      tags: [admin]
      x-actors: [admin]
      parameters: [{ in: query, name: status, schema: { $ref: '#/components/schemas/AppealStatus' } }, { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }]
      responses: { '200': { $ref: '#/components/responses/Success' }, '403': { $ref: '#/components/responses/Error' } }
  /admin/appeals/{id}/decision:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      operationId: decideAdminAppeal
      tags: [admin]
      x-actors: [admin]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AppealDecision' } } } }
      responses: { '200': { $ref: '#/components/responses/Success' }, '403': { $ref: '#/components/responses/Error' }, '404': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
...
components:
  parameters:
    Id: { in: path, name: id, required: true, schema: { type: string, minLength: 1 } }
...
    Page: { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
    Limit: { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
...
  responses:
    Success:
      description: Successful response envelope.
      content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
    Created:
      description: Resource created.
      content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
    Accepted:
      description: Request persisted for asynchronous delivery/processing.
      content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
    NoContent: { description: Mutation succeeded with no response body. }
    Error:
      description: Standardized application error; health and provider webhooks are exceptions.
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
  schemas:
    SuccessEnvelope: { $ref: './schemas/dtos.schema.json#/$defs/successEnvelope' }
...
    ErrorEnvelope: { $ref: './schemas/dtos.schema.json#/$defs/errorEnvelope' }
...
    ProfileWrite: { $ref: './schemas/dtos.schema.json#/$defs/profileWrite' }
    ChangePassword: { $ref: './schemas/dtos.schema.json#/$defs/changePassword' }
    EmailChange: { $ref: './schemas/dtos.schema.json#/$defs/emailChange' }
    EmailChangeVerify: { $ref: './schemas/dtos.schema.json#/$defs/emailChangeVerify' }
...
    UserRegister: { $ref: './schemas/dtos.schema.json#/$defs/userRegister' }
...
    AppealStatus: { $ref: './schemas/enums.schema.json#/$defs/appealStatus' }
...
    Role: { $ref: './schemas/enums.schema.json#/$defs/role' }
    UserStatus: { $ref: './schemas/enums.schema.json#/$defs/userStatus' }
...
    UserInvite: { $ref: './schemas/dtos.schema.json#/$defs/userInvite' }
    InvitationAccept: { $ref: './schemas/dtos.schema.json#/$defs/invitationAccept' }
    AdminUserWrite: { $ref: './schemas/dtos.schema.json#/$defs/adminUserWrite' }
    UserStatusWrite: { $ref: './schemas/dtos.schema.json#/$defs/userStatusWrite' }
    UserRoleWrite: { $ref: './schemas/dtos.schema.json#/$defs/userRoleWrite' }
    AdminPasswordResetRequest: { $ref: './schemas/dtos.schema.json#/$defs/adminPasswordResetRequest' }
    AppealDecision: { $ref: './schemas/dtos.schema.json#/$defs/appealDecision' }
...